Module Geojsonm

A library for manipulating large GeoJson documents without reading the whole document into memory using the Jsonm streaming, JSON parser.

module Err : sig ... end

type geometry =

| Point

| MultiPoint

| LineString

| MultiLineString

| Polygon

| MultiPolygon	(*
Types for geometry objects to be used in mapping over coordinates (hence the GeometryCollection is absent).
*)

type document =

| FeatureCollection

| Feature

| Geometry	(*
A GeoJson text is a single object of one of these types.
*)

Maps

Maps are functions that allow you to manipulate common structure in GeoJson objects. These will be written directly back to the destination that you provide.

For example, you might want to scale all coordinates by a factor of two.

let scale factor decoder = map_coords decoder (Array.map (( *. ) factor))

val map_coords : Jsonm.src -> Jsonm.dst -> f:(geometry -> float array -> float array) -> unit

map_coords src dst ~f will apply f to all GeoJson objects with something coordinate-like. This is essentially any geometry object. The type of geometry object currently being mapped over is also provided in case you wish to be more specific.

The map will recurse into geometry collections.

val map_props : Jsonm.src -> Jsonm.dst -> f:(Ezjsonm.value -> Ezjsonm.value) -> (unit, Err.t) Stdlib.result

map_props src dst ~f will apply f to each feature's properties field. The properties field is decoded into an Ezjsonm.value for convenience.

Folds

Folds are like maps except you can collect items into an accumulator which is returned to you.

For example, you might want to collect all of the names in the properties of features.

let get_string_exn = function `String s -> s | _ -> failwith "err"

let get_name = function
  | `O assoc -> List.assoc "name" assoc |> get_string_exn
  | _ -> failwith "err"

let places src =
  Geojsonm.fold_props src ~f:(fun acc p -> get_name p :: acc) []

val fold_props : Jsonm.src -> f:('a -> Ezjsonm.value -> 'a) -> init:'a -> ('a, Err.t) Stdlib.result

fold_props src f init